Battery check done!

+++ sensor/src/main/java/com/se459/sensor/interfaces/ICell.java
@@ -4,10 +4,10 @@ import com.se459.sensor.enums.PathType;
 import com.se459.sensor.enums.SurfaceType;
 
 public interface ICell {
-	public static final int UNKNOWN_COST = 3;
-	public static final int LOWPILE_COST = 2;
-	public static final int HIGHPILE_COST = 3;
-	public static final int BAREFLOOR_COST = 1;
+	public static final double UNKNOWN_COST = 3;
+	public static final double LOWPILE_COST = 2;
+	public static final double HIGHPILE_COST = 3;
+	public static final double BAREFLOOR_COST = 1;
 	int getX();
 	int getY();
 	SurfaceType getSurfaceType();
@@ -17,8 +17,8 @@ public interface ICell {
 	PathType getPathPosY();
 	PathType getPathNegY();
 	boolean getIsChargingStation();
-	int getTraverseCost();
-	int getVacuumCost();
+	double getTraverseCost();
+	double getVacuumCost();
 	void cleanCell();
 	void setTraversed();
 	boolean isTraversed();
